[增加]1. 增加消息错误码写入的转换

+++ GameFrameX.Hotfix/Player/Role/Role/Agent/RoleComponentAgent.cs
@@ -88,7 +88,7 @@ namespace GameFrameX.Hotfix.Player.Role.Role.Agent
             var channel = SessionManager.GetChannel(ActorId);
             if (channel != null && !channel.IsClose())
             {
-                channel.WriteAsync(msg, uniId, code);
+                channel.WriteAsync(msg, uniId, (int)code);
             }
         }
     }
